FULL

+++ cart.php
@@ -1,8 +1,11 @@
 <?php
 // header import
 include('./header.php');
+
 if (isset($_POST['delete-cart-submit'])) {
-    $deletedrecord = $cart->deleteCart($_POST['id']);
+    if ($products->getProducts('cart')[0]['user_id'] == $_SESSION['user']['id']) {
+        $deletedrecord = $cart->deleteCart($_POST['id']);
+    }
 }
 ?>
 
@@ -23,8 +26,9 @@ if (isset($_POST['delete-cart-submit'])) {
             </div>
         </div>
     </section>
-
+    
     <section class="shoping-cart spad">
+        <?php if (isset($_SESSION['user']) && $products->getProducts('cart')) {?>
         <div class="container">
             <div class="row">
                 <div class="col-lg-12">
@@ -41,39 +45,43 @@ if (isset($_POST['delete-cart-submit'])) {
                             </thead>
                             <tbody>
                                 <?php
-                                foreach ($products->getProducts('cart') as $item) :
-                                    $cart = $products->getProduct($item['item_id']);
-                                    $subTotal[] = array_map(function ($item) {
+                                    foreach ($products->getProducts('cart') as $item) :
+                                        if ($item['user_id'] == $_SESSION['user']['id']) {
+                                            $cart = $products->getProduct($item['item_id']);
+                                            $subTotal[] = array_map(function ($item) {
                                 ?>
-                                        <tr>
-                                            <td class="shoping__cart__item">
-                                                <img src="./img/<?php echo $item['image']; ?>" alt="">
-                                                <h5><?php echo $item['name'] ?? "Unknown"; ?></h5>
-                                            </td>
-                                            <td class="shoping__cart__price">
-                                                $<?php echo $item['price'] ?? 0; ?>
-                                            </td>
-                                            <td class="shoping__cart__quantity">
-                                                <div class="quantity">
-                                                    <div class="pro-qty">
-                                                        <button class="qty-up border bg-light" data-id="<?php echo $item['id'] ?? 0 ?>"><i class="fas fa-angle-up"></i></button>
-                                                        <input type="text" data-id="<?php echo $item['id'] ?? 0 ?>" class="qty_input border px-2 w-100 bg-light" disabled value="1" placeholder="1">
-                                                        <button data-id="<?php echo $item['id'] ?? 0 ?>" class="qty-down border bg-light"><i class="fas fa-angle-down"></i></button>
-                                                    </div>
-                                                </div>
-                                            </td>
-                                            <td class="shoping__cart__total">
-                                                    <form method="post">
-                                                        <input type="hidden" value="<?php echo $item['id'] ?? 0; ?>" name="id">
-                                                        <button type="submit" name="delete-cart-submit" class="btn font-baloo text-danger px-3 border-right">Delete</button>
-                                                    </form>
-                                            </td>
-                                        </tr>
+                                                <tr>
+                                                    <td class="shoping__cart__item">
+                                                        <img src="./img/<?php echo $item['image']; ?>" alt="">
+                                                        <h5><?php echo $item['name'] ?? "Unknown"; ?></h5>
+                                                    </td>
+                                                    <td class="shoping__cart__price">
+                                                        $<?php echo $item['price'] ?? 0; ?>
+                                                    </td>
+                                                    <td class="shoping__cart__quantity">
+                                                        <div class="quantity">
+                                                            <div class="pro-qty">
+                                                                <button class="qty-up border bg-light" data-id="<?php echo $item['id'] ?? 0 ?>"><i class="fas fa-angle-up"></i></button>
+                                                                <input type="text" data-id="<?php echo $item['id'] ?? 0 ?>" class="qty_input border px-2 w-100 bg-light" disabled value="1" placeholder="1">
+                                                                <button data-id="<?php echo $item['id'] ?? 0 ?>" class="qty-down border bg-light"><i class="fas fa-angle-down"></i></button>
+                                                            </div>
+                                                        </div>
+                                                    </td>
+                                                    <td class="shoping__cart__total">
+                                                        <form method="post">
+                                                            <input type="hidden" value="<?php echo $item['id'] ?? 0; ?>" name="id">
+                                                            <button type="submit" name="delete-cart-submit" class="btn font-baloo text-danger px-3 border-right">Delete</button>
+                                                        </form>
+                                                    </td>
+                                                </tr>
                                 <?php
-                                        return $item['price'];
-                                    }, $cart); // closing array_map function
-                                endforeach;
+                                                return $item['price'];
+                                            }, $cart); // closing array_map function
+                                        }
+                                    endforeach;
                                 ?>
+                                
+                                
                             </tbody>
                         </table>
                     </div>
@@ -101,7 +109,7 @@ if (isset($_POST['delete-cart-submit'])) {
                     <div class="shoping__checkout">
                         <h5>Cart Total</h5>
                         <div class="border-top py-4">
-                        <h5 class="font-baloo font-size-20">Subtotal ( <?php echo isset($subTotal) ? count($subTotal) : 0; ?> item):&nbsp; <span class="text-danger">$<span class="text-danger" id="deal-price"><?php echo isset($subTotal) ? $cart->getSum($subTotal) : 0; ?></span> </span> </h5>
+                            <h5 class="font-baloo font-size-20">Subtotal ( <?php echo isset($subTotal) ? count($subTotal) : 0; ?> item):&nbsp; <span class="text-danger">$<span class="text-danger" id="deal-price"><?php echo isset($subTotal) ? $cart->getSum($subTotal) : 0; ?></span> </span> </h5>
                             <button type="submit" class="btn btn-warning mt-3">Proceed to Buy</button>
                         </div>
                         <a href="#" class="primary-btn">PROCEED TO CHECKOUT</a>
@@ -109,6 +117,9 @@ if (isset($_POST['delete-cart-submit'])) {
                 </div>
             </div>
         </div>
+        <?php }else {
+            echo "<h3 class='text-center'>Cart is empty yet :(</h3>";
+        } ?>
     </section>
 </main>
